requirements.txt
openai>=1.60.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.10.0
fastapi>=0.115.0
uvicorn>=0.30.0
